+++ py/SGA/util.py
@@ -1,0 +1,44 @@
+"""
+SGA.util
+========
+
+Support utilities.
+
+"""
+import numpy as np
+
+def weighted_partition(weights, n):
+    '''
+    Partition `weights` into `n` groups with approximately same sum(weights)
+
+    Args:
+        weights: array-like weights
+        n: number of groups
+
+    Returns list of lists of indices of weights for each group
+
+    Notes:
+        compared to `dist_discrete_all`, this function allows non-contiguous
+        items to be grouped together which allows better balancing.
+
+    '''
+    #- sumweights will track the sum of the weights that have been assigned
+    #- to each group so far
+    sumweights = np.zeros(n, dtype=float)
+
+    #- Initialize list of lists of indices for each group
+    groups = list()
+    for i in range(n):
+        groups.append(list())
+
+    #- Assign items from highest weight to lowest weight, always assigning
+    #- to whichever group currently has the fewest weights
+    weights = np.asarray(weights)
+    for i in np.argsort(-weights):
+        j = np.argmin(sumweights)
+        groups[j].append(i)
+        sumweights[j] += weights[i]
+
+    assert len(groups) == n
+
+    return groups
